requirements.txt
python-socketio
eventlet
aiohttp
PyMuPDF
requests
